Added mipmaps, AO and performance optimizations
- Implemented post-processing shaders
- Created vertex and fragment shaders for preview rendering
- Added translucent shaders for rendering semi-transparent blocks
- New console system with commands
- Added a logger class for improved logging
- 16 new glass colors
- New parameter for blocks "translucent"
- Increased amount of possible IDs for blocks from 255 to 65535 (from uint8_t to uint16_t)
- Model can now use both cuboids and planes
- Implemented ambient occlusion
- More settings in-game and in options.txt
- Multiple performance optimizations regarding chunk generation, and face sorting
- Changed how models define UVs
- Added fallback block, biome and model
- Added underwater movement physics
- Switched from texture atlas to texture array
- Made camera block step smoother
- Fullscreen mode

+++ src/core/saveManager.hpp
@@ -24,8 +24,8 @@ public:
     static int getActiveSeed();
     static const WorldInfo* getActiveWorld();
     static void clearActiveWorld();
-    static bool savePlayerState(const glm::dvec3& position, float yaw, float pitch, const std::array<uint8_t, 9>& hotbar, bool flyEnabled);
-    static bool loadPlayerState(glm::dvec3& position, float& yaw, float& pitch, std::array<uint8_t, 9>& hotbar, bool& flyEnabled);
+    static bool savePlayerState(const glm::dvec3& position, float yaw, float pitch, const std::array<uint16_t, 9>& hotbar, bool flyEnabled);
+    static bool loadPlayerState(glm::dvec3& position, float& yaw, float& pitch, std::array<uint16_t, 9>& hotbar, bool& flyEnabled);
 
 private:
     static std::string generateUUID();
