+++ src/renderer/blockPreviewRenderer.hpp
@@ -9,7 +9,7 @@ class BlockPreviewRenderer {
 public:
     static void init(GLuint textureAtlas);
     static void generatePreviews();
-    static GLuint getPreviewTexture(uint8_t blockId);
+    static GLuint getPreviewTexture(uint16_t blockId);
     static void cleanup();
 
 private:
@@ -17,8 +17,8 @@ private:
     static GLuint shaderProgram;
     static GLuint fbo;
     static GLuint depthRbo;
-    static std::unordered_map<uint8_t, GLuint> previewTextures;
+    static std::unordered_map<uint16_t, GLuint> previewTextures;
 
     static GLuint createPreviewShader();
-    static void buildBlockMesh(uint8_t blockId, std::vector<float>& vertices, std::vector<unsigned int>& indices);
+    static void buildBlockMesh(uint16_t blockId, std::vector<float>& vertices, std::vector<unsigned int>& indices);
 };
